Fix #12: 'raw' data transform was not being found
Integration tests were failing when run after others in the browser. After
running with the PhantomJS runner, I saw more stack traces. All the failures
indicated that the 'raw' transform was not being found by Ember. I noticed
that some of the magic naming might have been inconsistent--taking a guess at
what the magic needed to be is the fix here. Why this wasn't a problem using
the app in a browser, I have no idea. Ugh.